refactor: move Nova wrapper into bedrock_models and share helpers
Per review, NovaMultimodalEmbeddingsModel now lives alongside BedrockModel
instead of in its own file.

Extracts three pieces both classes use:
- get_bedrock_runtime_client() for client construction
- CHARS_PER_TOKEN for the pre-truncation heuristic
- read_response_body() for decoding InvokeModel responses

Nova stays a separate class rather than a third provider branch on
BedrockModel: BedrockModel.encode is text-only by construction, and Nova's
request body and response shape differ from Titan's.
