+++ mteb/models/model_implementations/bedrock_models.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
+import base64
+import io
 import json
 import logging
 import re
+from concurrent.futures import ThreadPoolExecutor
 from typing import TYPE_CHECKING, Any
 
 import numpy as np
@@ -19,6 +22,7 @@ from .cohere_models import (
 )
 
 if TYPE_CHECKING:
+    from PIL import Image
     from torch.utils.data import DataLoader
 
     from mteb.abstasks.task_metadata import TaskMetadata
@@ -26,6 +30,30 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger(__name__)
+
+# Rough characters-per-token ratio used to pre-truncate text before sending it
+# to Bedrock.
+# https://docs.aws.amazon.com/bedrock/latest/userguide/titan-embedding-models.html
+CHARS_PER_TOKEN = 4.5
+
+
+def get_bedrock_runtime_client(region_name: str | None = None, config: Any = None):
+    """Create a bedrock-runtime client.
+
+    Defaults to the region of the active boto3 session when none is given.
+    """
+    import boto3
+
+    if region_name is None:
+        region_name = boto3.session.Session().region_name
+    if config is None:
+        return boto3.client("bedrock-runtime", region_name)
+    return boto3.client("bedrock-runtime", region_name, config=config)
+
+
+def read_response_body(response: Any) -> dict[str, Any]:
+    """Read and JSON-decode the streaming body of an InvokeModel response."""
+    return json.loads(response.get("body").read())
 
 
 class BedrockModel(AbsEncoder):
@@ -38,11 +66,7 @@ class BedrockModel(AbsEncoder):
         model_prompts: dict[str, str] | None = None,
         **kwargs,
     ) -> None:
-        import boto3
-
-        boto3_session = boto3.session.Session()
-        region_name = boto3_session.region_name
-        self._client = boto3.client("bedrock-runtime", region_name)
+        self._client = get_bedrock_runtime_client()
 
         self._model_id = model_id
         self._provider = provider.lower()
@@ -87,8 +111,7 @@ class BedrockModel(AbsEncoder):
         from botocore.exceptions import ValidationError
 
         all_embeddings = []
-        # https://docs.aws.amazon.com/bedrock/latest/userguide/titan-embedding-models.html
-        max_sequence_length = int(self._max_tokens * 4.5)
+        max_sequence_length = int(self._max_tokens * CHARS_PER_TOKEN)
 
         for sentence in tqdm(sentences, leave=False, disable=not show_progress_bar):
             if len(sentence) > max_sequence_length:
@@ -156,7 +179,7 @@ class BedrockModel(AbsEncoder):
         return self._to_numpy(response)
 
     def _to_numpy(self, embedding_response) -> Array:
-        response = json.loads(embedding_response.get("body").read())
+        response = read_response_body(embedding_response)
         key = "embedding" if self._provider == "amazon" else "embeddings"
         return np.array(response[key])
 
@@ -278,5 +301,163 @@ cohere_embed_multilingual_v3 = ModelMeta(
     similarity_fn_name=ScoringFunction.COSINE,
     framework=["API"],
     use_instructions=True,
+    extra_requirements_groups=["boto3"],
+)
+
+
+# embeddingPurpose values accepted by Nova Multimodal Embeddings.
+NOVA_PURPOSE_DOCUMENT = "GENERIC_INDEX"
+NOVA_PURPOSE_QUERY = "GENERIC_RETRIEVAL"
+NOVA_PURPOSE_CLASSIFICATION = "CLASSIFICATION"
+NOVA_PURPOSE_CLUSTERING = "CLUSTERING"
+
+_NOVA_CLASSIFICATION_TASKS = {
+    "Classification",
+    "MultilabelClassification",
+    "PairClassification",
+}
+
+
+class NovaMultimodalEmbeddingsModel(AbsEncoder):
+    """Amazon Nova Multimodal Embeddings via the Bedrock synchronous InvokeModel API.
+
+    Text and image only. Nova also supports audio and video, but mteb decodes
+    both before they reach the encoder (audio as a float array, video as a
+    torchcodec frame tensor), so sending them would require re-encoding decoded
+    media back into a container.
+    """
+
+    def __init__(
+        self,
+        model_id: str = "amazon.nova-2-multimodal-embeddings-v1:0",
+        region_name: str = "us-east-1",
+        embedding_dimension: int = 3072,
+        max_tokens: int = 8192,
+        max_workers: int = 8,
+        **kwargs: Any,
+    ) -> None:
+        from botocore.config import Config
+
+        self._model_id = model_id
+        self._embedding_dimension = embedding_dimension
+        self._max_workers = max_workers
+        self._max_chars = int(max_tokens * CHARS_PER_TOKEN)
+        self._client = get_bedrock_runtime_client(
+            region_name,
+            config=Config(retries={"max_attempts": 10, "mode": "adaptive"}),
+        )
+
+    @staticmethod
+    def _purpose(task_metadata: TaskMetadata, prompt_type: PromptType | None) -> str:
+        if task_metadata.type in _NOVA_CLASSIFICATION_TASKS:
+            return NOVA_PURPOSE_CLASSIFICATION
+        if task_metadata.type == "Clustering":
+            return NOVA_PURPOSE_CLUSTERING
+        if (
+            prompt_type is not None
+            and getattr(prompt_type, "value", prompt_type) == "query"
+        ):
+            return NOVA_PURPOSE_QUERY
+        return NOVA_PURPOSE_DOCUMENT
+
+    def _params(self, purpose: str) -> dict[str, Any]:
+        return {
+            "embeddingPurpose": purpose,
+            "embeddingDimension": self._embedding_dimension,
+        }
+
+    def _text_payload(self, text: str, purpose: str) -> dict[str, Any]:
+        params = self._params(purpose)
+        params["text"] = {
+            "truncationMode": "END",
+            "value": text[: self._max_chars],
+        }
+        return {"taskType": "SINGLE_EMBEDDING", "singleEmbeddingParams": params}
+
+    def _image_payload(self, image: Image.Image, purpose: str) -> dict[str, Any]:
+        buffer = io.BytesIO()
+        image.convert("RGB").save(buffer, format="JPEG", quality=95)
+        params = self._params(purpose)
+        params["image"] = {
+            "format": "jpeg",
+            "source": {"bytes": base64.b64encode(buffer.getvalue()).decode()},
+        }
+        return {"taskType": "SINGLE_EMBEDDING", "singleEmbeddingParams": params}
+
+    def _invoke(self, payload: dict[str, Any]) -> list[float]:
+        response = self._client.invoke_model(
+            modelId=self._model_id, body=json.dumps(payload)
+        )
+        return read_response_body(response)["embeddings"][0]["embedding"]
+
+    def encode(
+        self,
+        inputs: DataLoader[BatchedInput],
+        *,
+        task_metadata: TaskMetadata,
+        hf_split: str,
+        hf_subset: str,
+        prompt_type: PromptType | None = None,
+        **kwargs: Any,
+    ) -> Array:
+        purpose = self._purpose(task_metadata, prompt_type)
+        show_progress_bar = kwargs.pop("show_progress_bar", False)
+
+        embeddings: list[list[float]] = []
+        pbar = tqdm(disable=not show_progress_bar, leave=False, desc="Nova MME")
+
+        with ThreadPoolExecutor(max_workers=self._max_workers) as pool:
+            for batch in inputs:
+                images = batch.get("image")
+                texts = batch.get("text")
+
+                if images and texts:
+                    raise NotImplementedError(
+                        "Interleaved text+image input is not yet supported for "
+                        "Nova Multimodal Embeddings."
+                    )
+                if images:
+                    payloads = [self._image_payload(im, purpose) for im in images]
+                elif texts:
+                    payloads = [self._text_payload(t, purpose) for t in texts]
+                else:
+                    raise ValueError(
+                        f"Batch has no 'text' or 'image' key. Got: {list(batch.keys())}"
+                    )
+
+                embeddings.extend(pool.map(self._invoke, payloads))
+                pbar.update(len(payloads))
+
+        pbar.close()
+        return np.array(embeddings, dtype=np.float32)
+
+
+amazon_nova_2_multimodal_embeddings = ModelMeta(
+    name="bedrock/amazon-nova-2-multimodal-embeddings",
+    model_type=["dense"],
+    revision="1",
+    release_date="2025-10-28",
+    languages=None,  # 200+ languages, not enumerated by the provider
+    loader=NovaMultimodalEmbeddingsModel,
+    loader_kwargs=dict(
+        model_id="amazon.nova-2-multimodal-embeddings-v1:0",
+        embedding_dimension=3072,
+        max_tokens=8192,
+    ),
+    max_tokens=8192,
+    embed_dim=3072,
+    open_weights=False,
+    n_parameters=None,
+    n_embedding_parameters=None,
+    memory_usage_mb=None,
+    public_training_code=None,
+    public_training_data=None,
+    training_datasets=None,
+    license=None,
+    reference="https://docs.aws.amazon.com/nova/latest/userguide/nova-embeddings.html",
+    similarity_fn_name=ScoringFunction.COSINE,
+    framework=["API"],
+    use_instructions=False,
+    modalities=["text", "image"],
     extra_requirements_groups=["boto3"],
 )
